% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avro.R
\name{avro.array}
\alias{avro.array}
\title{avro.array}
\usage{
avro.array(items)
}
\arguments{
\item{items}{schema for the homogeneous array}
}
\description{
Constructs a list-of-lists Avro schema for the array type.
}
\examples{
avro.array(avro.int)
avro.array(avro.string)
}

