% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avro.R
\docType{data}
\name{avro.bytes}
\alias{avro.bytes}
\title{avro.bytes}
\format{An object of class \code{character} of length 1.}
\usage{
avro.bytes
}
\description{
Constructs a list-of-lists Avro schema for the bytes (unstructured byte array) type.
}
\keyword{datasets}

