% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/avro.R
\docType{data}
\name{avro.double}
\alias{avro.double}
\title{avro.double}
\format{\preformatted{ chr "double"
}}
\usage{
avro.double
}
\description{
Constructs a list-of-lists Avro schema for the double (floating-point numeric with 64-bit precision) type.
}
\keyword{datasets}

