% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avro.R
\name{avro.enum}
\alias{avro.enum}
\title{avro.enum}
\usage{
avro.enum(symbols, name = NULL, namespace = NULL)
}
\arguments{
\item{symbols}{list of string-valued symbol names}

\item{name}{required name (if missing, uniqueEnumName is invoked)}

\item{namespace}{optional namespace}
}
\description{
Constructs a list-of-lists Avro schema for the enum (set of symbols) type.
}
\examples{
avro.enum(list("one", "two", "three"))
}

