% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avro.R
\name{avro.fixed}
\alias{avro.fixed}
\title{avro.fixed}
\usage{
avro.fixed(size, name = NULL, namespace = NULL)
}
\arguments{
\item{size}{size of the byte array}

\item{name}{required name (if missing, uniqueFixedName is invoked)}

\item{namespace}{optional namespace}
}
\description{
Constructs a list-of-lists Avro schema for the fixed (byte array with fixed size) type.
}
\examples{
avro.fixed(6, "MACAddress")
}

