% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avro.R
\docType{data}
\name{avro.float}
\alias{avro.float}
\title{avro.float}
\format{An object of class \code{character} of length 1.}
\usage{
avro.float
}
\description{
Constructs a list-of-lists Avro schema for the float (floating-point numeric with 32-bit precision) type.
}
\keyword{datasets}

