% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avro.fromFrame.R
\name{avro.fromFrame}
\alias{avro.fromFrame}
\title{avro.fromFrame}
\usage{
avro.fromFrame(dataFrame, exclude = list(), name = NULL, namespace = NULL)
}
\arguments{
\item{dataFrame}{a data.frame object}

\item{exclude}{set of field names to exclude}

\item{name}{required name of the record (if not specified, uniqueRecordName will be invoked)}

\item{namespace}{optional namespace of the record}
}
\value{
Avro list-of-lists representing a record type
}
\description{
Convenience function for creating an Avro input schema from a data frame
}
\examples{
avro.fromFrame(data.frame(xx = c(1, 3, 5)))
}

