% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avro.fullName.R
\name{avro.fullName}
\alias{avro.fullName}
\title{avro.fullName}
\usage{
avro.fullName(type)
}
\arguments{
\item{type}{Avro list-of-lists}
}
\value{
string representing the full name
}
\description{
Yields the full type name (with namespace) of an Avro list-of-lists
}
\examples{
avro.fullName(avro.record(list(), "MyRecord"))                   # "MyRecord"
avro.fullName(avro.record(list(), "MyRecord", "com.wowzers"))    # "com.wowzers.MyRecord"
}

