% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avro.R
\docType{data}
\name{avro.long}
\alias{avro.long}
\title{avro.long}
\format{An object of class \code{character} of length 1.}
\usage{
avro.long
}
\description{
Constructs a list-of-lists Avro schema for the long (integer numeric with 64-bit precision) type.
}
\keyword{datasets}

