% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/avro.R
\name{avro.map}
\alias{avro.map}
\title{avro.map}
\usage{
avro.map(values)
}
\arguments{
\item{values}{schema for the homogeneous map}
}
\description{
Constructs a list-of-lists Avro schema for the map type.
}
\examples{
avro.map(avro.int)
avro.map(avro.string)
}

