% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/avro.R
\name{avro.record}
\alias{avro.record}
\title{avro.record}
\usage{
avro.record(fields, name = NULL, namespace = NULL)
}
\arguments{
\item{fields}{named list of field names and schemas}

\item{name}{required name (if missing, uniqueRecordName is invoked)}

\item{namespace}{optional namespace}
}
\description{
Constructs a list-of-lists Avro schema for the record type.
}
\examples{
avro.record(list(one = avro.int, two = avro.double, three = avro.string))
}

