% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/avro.R
\docType{data}
\name{avro.string}
\alias{avro.string}
\title{avro.string}
\format{\preformatted{ chr "string"
}}
\usage{
avro.string
}
\description{
Constructs a list-of-lists Avro schema for the string (UTF-8) type.
}
\keyword{datasets}

