% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avro.type.R
\name{avro.type}
\alias{avro.type}
\title{avro.type}
\usage{
avro.type(obj)
}
\arguments{
\item{obj}{object to inspect}
}
\value{
list-of-lists Avro schema
}
\description{
Inspects an R object and produces the corresponding Avro type name
}
\examples{
avro.type("hello")           # "string"
avro.type(factor("hello"))   # "string"
avro.type(3.14)              # "double"
avro.type(3)                 # "int"
}

