% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avro.R
\name{avro.union}
\alias{avro.union}
\title{avro.union}
\usage{
avro.union(...)
}
\arguments{
\item{...}{schemas for each of the possible sub-types}
}
\description{
Constructs a list-of-lists Avro schema for the tagged union type.
}
\examples{
avro.union(avro.null, avro.int)         # a way to make a nullable int
avro.union(avro.double, avro.string)    # any set of types can be unioned
}

