% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avro.R
\name{avro_array}
\alias{avro_array}
\title{avro_array}
\usage{
avro_array(items)
}
\arguments{
\item{items}{schema for the homogeneous array}
}
\description{
Constructs a \code{list} of lists Avro schema for the array type.
}
\examples{
avro_array(avro_int)
avro_array(avro_string)
}
