% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avro.R
\docType{data}
\name{avro_boolean}
\alias{avro_boolean}
\title{avro_boolean}
\format{An object of class \code{character} of length 1.}
\usage{
avro_boolean
}
\description{
Constructs a \code{list} of lists Avro schema for the boolean (logical) type.
}
\keyword{datasets}
