% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avro.R
\docType{data}
\name{avro_bytes}
\alias{avro_bytes}
\title{avro_bytes}
\format{An object of class \code{character} of length 1.}
\usage{
avro_bytes
}
\description{
Constructs a \code{list} of lists Avro schema for the bytes (unstructured byte array) type.
}
\keyword{datasets}
