% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avro.R
\docType{data}
\name{avro_double}
\alias{avro_double}
\title{avro_double}
\format{An object of class \code{character} of length 1.}
\usage{
avro_double
}
\description{
Constructs a \code{list} of lists Avro schema for the double (floating-point numeric with 64-bit precision) type.
}
\keyword{datasets}
