% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avro.R
\name{avro_enum}
\alias{avro_enum}
\title{avro_enum}
\source{
gen_unique_name.R
}
\usage{
avro_enum(symbols, name = NULL, namespace = NULL)
}
\arguments{
\item{symbols}{list of string-valued symbol names}

\item{name}{required name (if missing, \code{\link{gen_unique_enum_name}} is invoked)}

\item{namespace}{optional namespace}
}
\description{
Constructs a \code{list} of lists Avro schema for the enum (set of symbols) type.
}
\examples{
avro_enum(list("one", "two", "three"))
}
