% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avro.R
\docType{data}
\name{avro_float}
\alias{avro_float}
\title{avro_float}
\format{An object of class \code{character} of length 1.}
\usage{
avro_float
}
\description{
Constructs a \code{list} of lists Avro schema for the float (floating-point numeric with 32-bit precision) type.
}
\keyword{datasets}
