% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avro_from_df.R
\name{avro_from_df}
\alias{avro_from_df}
\title{avro_from_df}
\usage{
avro_from_df(df, exclude = list(), name = NULL, namespace = NULL)
}
\arguments{
\item{df}{a \code{data.frame}}

\item{exclude}{set of field names to exclude}

\item{name}{required name of the record (if not specified, gen_unique_rec_name() will be invoked)}

\item{namespace}{optional namespace of the record}
}
\value{
a \code{list} of lists representing an Avro record type
}
\description{
Convenience function for creating an Avro input schema from a data frame
}
\examples{
avro_from_df(data.frame(x = c(1, 3, 5)))
}
