% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avro_fullname.R
\name{avro_fullname}
\alias{avro_fullname}
\title{avro_fullname}
\usage{
avro_fullname(type)
}
\arguments{
\item{type}{Avro \code{list} of lists}
}
\value{
string representing the full name
}
\description{
Yields the full type name (with namespace) of an Avro \code{list} of lists
}
\examples{
avro_fullname(avro_record(list(), "MyRecord"))                   # "MyRecord"
avro_fullname(avro_record(list(), "MyRecord", "com.wowzers"))    # "com.wowzers.MyRecord"
}
