% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avro.R
\docType{data}
\name{avro_int}
\alias{avro_int}
\title{avro_int}
\format{An object of class \code{character} of length 1.}
\usage{
avro_int
}
\description{
Constructs a \code{list} of lists Avro schema for the int (integer numeric with 32-bit precision) type.
}
\keyword{datasets}
