% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avro.R
\name{avro_map}
\alias{avro_map}
\title{avro_map}
\usage{
avro_map(values)
}
\arguments{
\item{values}{schema for the homogeneous map}
}
\description{
Constructs a \code{list} of lists Avro schema for the map type.
}
\examples{
avro_map(avro_int)
avro_map(avro_string)
}
