% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avro.R
\docType{data}
\name{avro_null}
\alias{avro_null}
\title{avro_null}
\format{An object of class \code{character} of length 1.}
\usage{
avro_null
}
\description{
Constructs a \code{list} of lists Avro schema for the null type (type with only one value).
}
\keyword{datasets}
