% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avro.R
\name{avro_record}
\alias{avro_record}
\title{avro_record}
\usage{
avro_record(fields, name = NULL, namespace = NULL)
}
\arguments{
\item{fields}{named list of field names and schemas}

\item{name}{required name (if missing, gen_unique_rec_name is invoked)}

\item{namespace}{optional namespace}
}
\description{
Constructs a \code{list} of lists Avro schema for the record type.
}
\examples{
avro_record(list(one = avro_int, two = avro_double, three = avro_string))
}
