% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avro.R
\docType{data}
\name{avro_string}
\alias{avro_string}
\title{avro_string}
\format{An object of class \code{character} of length 1.}
\usage{
avro_string
}
\description{
Constructs a \code{list} of lists Avro schema for the string (UTF-8) type.
}
\keyword{datasets}
