% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avro_type.R
\name{avro_type}
\alias{avro_type}
\title{avro_type}
\usage{
avro_type(obj)
}
\arguments{
\item{obj}{object to inspect}
}
\value{
a \code{list} of lists Avro schema
}
\description{
Inspects an R object and produces the corresponding Avro type name
}
\examples{
avro_type("hello")           # "string"
avro_type(factor("hello"))   # "string"
avro_type(3.14)              # "double"
avro_type(3)                 # "int"
}
