% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avro.R
\name{avro_union}
\alias{avro_union}
\title{avro_union}
\usage{
avro_union(...)
}
\arguments{
\item{...}{schemas for each of the possible sub-types}
}
\description{
Constructs a \code{list} of lists Avro schema for the tagged union type.
}
\examples{
avro_union(avro_null, avro_int)         # a way to make a nullable int
avro_union(avro_double, avro_string)    # any set of types can be unioned
}
