% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lda.R
\name{build_model.lda}
\alias{build_model.lda}
\title{build_model.lda}
\usage{
\method{build_model}{lda}(object, ...)
}
\arguments{
\item{object}{a object of class lda}

\item{...}{further arguments passed to or from other methods}
}
\value{
a \code{list} of lists representation of the linear discriminant model that can be
inserted into a cell or pool
}
\description{
Builds an entire PFA list of lists based on a lda() fit
}
\examples{
model <- MASS::lda(Species ~ ., data=iris)
model_built <- build_model(model)
}
