% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pfa.R
\name{build_model}
\alias{build_model}
\title{build_model}
\usage{
build_model(object, ...)
}
\arguments{
\item{object}{a model object}

\item{...}{further arguments passed to or from other methods}
}
\value{
a \code{list} of lists representation of the tree that can be 
inserted into a cell or pool
}
\description{
Builds an entire PFA list of lists based on a model object
}
\examples{
# all the "build_model" methods found
methods("build_model")
}
