% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpart.R
\name{build_model.rpart}
\alias{build_model.rpart}
\title{build_model.rpart}
\usage{
\method{build_model}{rpart}(object, ...)
}
\arguments{
\item{object}{a object of class "rpart"}

\item{...}{further arguments passed to or from other methods}
}
\value{
a \code{list} of lists representation of the tree that can be 
inserted into a cell or pool
}
\description{
Builds an entire PFA list of lists based on a single tree
}
\examples{
model <- rpart::rpart(Kyphosis ~ Age + as.factor(Number), data = rpart::kyphosis)
my_tree <- build_model(model)
}
