% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbm.R
\name{build_node_gbm}
\alias{build_node_gbm}
\title{build_node_gbm}
\usage{
build_node_gbm(tree, categorical_lookup, whichNode, valueNeedsTag, dataLevels,
  fieldTypes)
}
\arguments{
\item{tree}{tree object}

\item{whichNode}{left or right node for categoricalLookup}

\item{valueNeedsTag}{flag for whether node needs label}

\item{dataLevels}{levels of data}

\item{fieldTypes}{type of fields}

\item{categoricalLookup}{splits used}
}
\value{
PFA as a list-of-lists that can be inserted into a cell or pool
}
\description{
Builds one node of a gbm model tree
}

