% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomForest.R
\name{build_node_randomForest}
\alias{build_node_randomForest}
\title{build_node_randomForest}
\usage{
build_node_randomForest(tree_table, leaf_val_type, whichNode, valueNeedsTag,
  dataLevels, fieldTypes = NULL)
}
\arguments{
\item{tree_table}{tree object}

\item{leaf_val_type}{a character representing an avro type when a value is 
returned at a leaf node}

\item{whichNode}{pointer to the target node to be built}

\item{valueNeedsTag}{flag for whether node needs label}

\item{dataLevels}{levels of data}

\item{fieldTypes}{type of fields}
}
\value{
PFA as a list-of-lists that can be inserted into a cell or pool
}
\description{
Builds one node of a randomForest model tree
}
\keyword{internal}
