% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbm.R
\name{build_tree.gbm}
\alias{build_tree.gbm}
\title{build_node}
\usage{
build_tree.gbm(object, which_tree = 1)
}
\arguments{
\item{object}{a object of class gbm}

\item{which_tree}{an integer indicating which single tree to build}
}
\value{
a \code{list} of lists representation of the tree that can be 
inserted into a cell or pool
}
\description{
Builds an entire PFA list of lists based on a single gbm model tree
}
\examples{
binomial_dat <- data.frame(X1 = runif(100), 
                           X2 = rnorm(100), 
                           Y = (rexp(100,5) + 5 * X1 - 4 * X2) > 0)

bernoulli_model <- gbm(Y ~ X1 + X2, 
                       data=binomial_dat, 
                       distribution = 'bernoulli')
  
my_tree <- build_tree(bernoulli_model, 1)
}

