% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomForest.R
\name{build_tree.randomForest}
\alias{build_tree.randomForest}
\title{build_tree}
\usage{
\method{build_tree}{randomForest}(object, which_tree = 1)
}
\arguments{
\item{object}{a object of class randomForest}

\item{which_tree}{an integer indicating which single tree to build}
}
\value{
a \code{list} of lists representation of the tree that can be 
inserted into a cell or pool
}
\description{
Builds an entire PFA list of lists based on a single randomForest model tree
}
\examples{
binomial_dat <- data.frame(X1 = runif(100), 
                           X2 = rnorm(100))
binomial_dat$Y <- factor((rexp(100,5) + 
                            5 * binomial_dat$X1 - 
                            4 * binomial_dat$X2) > 0)

bernoulli_model <- randomForest(Y ~ X1 + X2, data=binomial_dat)

my_tree <- build_tree.randomForest(bernoulli_model, 1)
}

