% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pfa.R
\name{build_tree}
\alias{build_tree}
\title{build_tree}
\usage{
build_tree(object, ...)
}
\arguments{
\item{object}{a tree model object}
}
\value{
a \code{list} of lists representation of the tree that can be 
inserted into a cell or pool
}
\description{
Builds an entire PFA list of lists based on a single tree model object
}
\examples{
\dontrun{
require(utils)

# all the "build_tree" methods found
methods("build_tree")
}
}

