% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/converters.R
\name{convert_fcn}
\alias{convert_fcn}
\title{convert_fcn}
\usage{
convert_fcn(name, args, symbols, cells, pools, fcns, env)
}
\description{
A function to convert R expressions to their PFA equivalent
}
\keyword{internal}
