% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ets.R
\name{extract_params.ets}
\alias{extract_params.ets}
\title{extract_params.ets}
\usage{
\method{extract_params}{ets}(object, ...)
}
\arguments{
\item{object}{an object of class "ets"}

\item{...}{further arguments passed to or from other methods}
}
\value{
PFA as a list-of-lists that can be inserted into a cell or pool
}
\description{
Extract model parameters from an Exponential smoothing state space model created 
using the ets() function from the forecast package.
}
\examples{
model <- forecast::ets(USAccDeaths, model="ZZZ")
extracted_model <- extract_params(model)
}
