% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.R
\name{extract_params.forecast}
\alias{extract_params.forecast}
\title{extract_params.forecast}
\usage{
\method{extract_params}{forecast}(object, ...)
}
\arguments{
\item{object}{an object of class "forecast"}

\item{...}{further arguments passed to or from other methods}
}
\value{
PFA as a list-of-lists that can be inserted into a cell or pool
}
\description{
Extract model parameters from a model with class "forecast" created using 
the forecast package
}
\examples{
model <- forecast::holt(airmiles)
extracted_model <- extract_params(model)
}
