% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/holtwinters.R
\name{extract_params.HoltWinters}
\alias{extract_params.HoltWinters}
\title{extract_params.HoltWinters}
\usage{
\method{extract_params}{HoltWinters}(object, ...)
}
\arguments{
\item{object}{an object of class "HoltWinters"}

\item{...}{further arguments passed to or from other methods}
}
\value{
PFA as a list-of-lists that can be inserted into a cell or pool
}
\description{
Extract Holt Winters model parameters from the stats library
}
\examples{
model <- HoltWinters(co2)
extracted_model <- extract_params(model)
}
