% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knn.R
\name{extract_params.ipredknn}
\alias{extract_params.ipredknn}
\title{extract_params.ipredknn}
\usage{
\method{extract_params}{ipredknn}(object, ...)
}
\arguments{
\item{object}{an object of class "ipredknn"}

\item{...}{further arguments passed to or from other methods}
}
\value{
PFA as a list-of-lists that can be inserted into a cell or pool
}
\description{
Extract K-nearest neighbor model parameters from a ipredknn object created by 
the ipred library
}
\examples{
iris2 <- iris
colnames(iris2) <- gsub('\\\\.', '_', colnames(iris2))
model <- ipred::ipredknn(Species ~ ., iris2)
params <- extract_params(model)
}
