% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kcca.R
\name{extract_params.kcca}
\alias{extract_params.kcca}
\title{extract_params.kcca}
\usage{
\method{extract_params}{kcca}(object, ...)
}
\arguments{
\item{object}{an object of class "kcca"}

\item{...}{further arguments passed to or from other methods}
}
\value{
PFA as a list-of-lists that can be inserted into a cell or pool
}
\description{
Extract K-centroids model parameters from a kcca object created by 
the flexclust library
}
\examples{
model <- flexclust::kcca(iris[,1:4], k = 3, family=flexclust::kccaFamily("kmeans"))
extracted_params <- extract_params(model)
}
