% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kcca.R
\name{extract_params.kccasimple}
\alias{extract_params.kccasimple}
\title{extract_params.kccasimple}
\usage{
\method{extract_params}{kccasimple}(object, ...)
}
\arguments{
\item{object}{an object of class "kccasimple"}

\item{...}{further arguments passed to or from other methods}
}
\value{
PFA as a list-of-lists that can be inserted into a cell or pool
}
\description{
Extract K-centroids model parameters from a kccasimple object created by 
the flexclust library
}
\examples{
model <- flexclust::kcca(iris[,1:4], k = 3, 
                         family=flexclust::kccaFamily("kmeans"), simple=TRUE)
extracted_params <- extract_params(model)
}
