% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmeans.R
\name{extract_params.kmeans}
\alias{extract_params.kmeans}
\title{extract_params.kmeans}
\source{
kcca.R
}
\usage{
\method{extract_params}{kmeans}(object, ...)
}
\arguments{
\item{object}{an object of class "kmeans"}

\item{...}{further arguments passed to or from other methods}
}
\value{
PFA as a list-of-lists that can be inserted into a cell or pool
}
\description{
Extract K-means model parameters from a kmeans object
}
\examples{
model <- kmeans(x=iris[, 1:2], centers=3)
extracted_params <- extract_params(model)
}
