% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knn.R
\name{extract_params.knn3}
\alias{extract_params.knn3}
\title{extract_params.knn3}
\source{
pfa_utils.R
}
\usage{
\method{extract_params}{knn3}(object, ...)
}
\arguments{
\item{object}{an object of class "knn3"}

\item{...}{further arguments passed to or from other methods}
}
\value{
PFA as a list-of-lists that can be inserted into a cell or pool
}
\description{
Extract K-nearest neighbor model parameters from a knn3 object created by 
the caret library
}
\examples{
iris2 <- iris
colnames(iris2) <- gsub('\\\\.', '_', colnames(iris2))
model <- caret::knn3(Species ~ ., iris2)
extracted_params <- extract_params(model)
}
