% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knn.R
\name{extract_params.knnreg}
\alias{extract_params.knnreg}
\title{extract_params.knnreg}
\usage{
\method{extract_params}{knnreg}(object, ...)
}
\arguments{
\item{object}{an object of class "knnreg"}

\item{...}{further arguments passed to or from other methods}
}
\value{
PFA as a list-of-lists that can be inserted into a cell or pool
}
\description{
Extract K-nearest neighbor model parameters from a knnreg object created by 
the caret library
}
\examples{
model <- caret::knnreg(mpg ~ cyl + hp + am + gear + carb, data = mtcars)
extracted_params <- extract_params(model)
}
