% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lda.R
\name{extract_params.lda}
\alias{extract_params.lda}
\title{extract_params.lda}
\usage{
\method{extract_params}{lda}(object, ...)
}
\arguments{
\item{object}{an object of class "lda"}

\item{...}{further arguments passed to or from other methods}
}
\value{
PFA as a list-of-lists that can be inserted into a cell or pool
}
\description{
Extract linear discriminant model parameters from a model created
by the lda() function
}
\examples{
model <- MASS::lda(Species ~ ., data=iris)
model_params <- extract_params(model)
}
