% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pfa.R
\name{extract_params}
\alias{extract_params}
\title{extract_params}
\usage{
extract_params(object, ...)
}
\arguments{
\item{object}{a model object}
}
\value{
a \code{list} that is extracted from the tree model object
}
\description{
Extracts parameters of a model object
}
\examples{
\dontrun{
require(utils)

# all the "extract_params" methods found
methods("extract_params")
}
}

