% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbm.R
\name{extract_tree.gbm}
\alias{extract_tree.gbm}
\title{extract_tree}
\usage{
\method{extract_tree}{gbm}(object, which_tree = 1)
}
\arguments{
\item{object}{an object of class "gbm"}

\item{which_tree}{the number of the tree to extract}
}
\value{
a \code{list} that is extracted from the gbm object
}
\description{
Extracts a tree from an ensemble made by the gbm function
}
\examples{
binomial_dat <- data.frame(X1 = runif(100), 
                           X2 = rnorm(100), 
                           Y = (rexp(100,5) + 5 * X1 - 4 * X2) > 0)

bernoulli_model <- gbm(Y ~ X1 + X2, 
                       data=binomial_dat, 
                       distribution = 'bernoulli')
  
my_tree <- extract_tree(bernoulli_model, 1)
}

