% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pfa.R
\name{extract_tree}
\alias{extract_tree}
\title{extract_tree}
\usage{
extract_tree(object, ...)
}
\arguments{
\item{object}{a tree model object}
}
\value{
a \code{list} that is extracted from the tree model object
}
\description{
Extracts a tree from a tree model object
}
\examples{
\dontrun{
require(utils)

# all the "extract_tree" methods found
methods("extract_tree")
}
}

