% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_unique_name.R
\name{gen_unique_enum_name}
\alias{gen_unique_enum_name}
\title{gen_unique_enum_name}
\usage{
gen_unique_enum_name()
}
\value{
name in the form "Enum_###"
}
\description{
Convenience or internal function for generating enum names; each call results in a new name.
}
\examples{
gen_unique_enum_name()
}
