% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_unique_name.R
\name{gen_unique_fixed_name}
\alias{gen_unique_fixed_name}
\title{gen_unique_fixed_name}
\usage{
gen_unique_fixed_name()
}
\value{
name in the form "Fixed_###"
}
\description{
Convenience or internal function for generating fixed names; each call results in a new name.
}
\examples{
gen_unique_fixed_name()
}
