% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_unique_name.R
\name{gen_unique_symb_name}
\alias{gen_unique_symb_name}
\title{gen_unique_symb_name}
\usage{
gen_unique_symb_name(symbols)
}
\arguments{
\item{symbols}{Symbols}
}
\value{
name in the form "tmp_###"
}
\description{
Convenience or internal function for generating symbol names; each call results in a new name.
}
\examples{
gen_unique_symb_name()
}
