% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/jsonobj.R
\name{json.array}
\alias{json.array}
\title{json.array}
\usage{
json.array(...)
}
\arguments{
\item{args}{optional contents of the unnamed list}
}
\value{
an unnamed list
}
\description{
Convenience function for making a (possibly empty) unnamed list, which converts to a JSON array.
}
\examples{
json.array()
json.array(1, TRUE, "THREE")
}

