% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jsonobj.R
\name{json.map}
\alias{json.map}
\title{json.map}
\usage{
json.map(...)
}
\arguments{
\item{...}{optional contents of the named list (as key-value pairs)}
}
\value{
a named list
}
\description{
Convenience function for making a (possibly empty) named list, which converts to a JSON object.
}
\examples{
json.map()
json.map(one = 1, two = TRUE, three = "THREE")
}

