% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/json.R
\name{json}
\alias{json}
\title{json}
\usage{
json(x, fileName = NULL, newline = TRUE, spaces = TRUE, sigfigs = NULL,
  stringsAsIs = FALSE)
}
\arguments{
\item{x}{The structure to convert.}

\item{fileName}{If NULL (default), return an in-memory JSON string; if a fileName string, write to a file without incurring in-memory overhead.}

\item{newline}{If TRUE (default), end the string/file with a newline.}

\item{spaces}{If TRUE (default), include spaces after commas and colons for readability.}

\item{sigfigs}{If NULL (default), represent numbers with full precision; if an integer, represent numbers with that many significant digits.}

\item{stringsAsIs}{If FALSE (default), process strings to escape characters that need to be escaped in JSON; if TRUE, pass strings as-is for speed, possibly creating invalid JSON.}
}
\value{
describeme
}
\description{
Convert a list-of-lists structure into a JSON string in memory or a JSON file on disk.
}
\examples{
cat(json(pfaDocument))
json(pfaDocument, fileName = "myModel.pfa")
}

