% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jsonobj.R
\name{json_array}
\alias{json_array}
\title{json_array}
\usage{
json_array(...)
}
\arguments{
\item{...}{optional contents of the unnamed list}
}
\value{
an unnamed list
}
\description{
Convenience function for making a (possibly empty) unnamed list, which 
converts to a JSON array.
}
\examples{
json_array()
json_array(1, TRUE, "THREE")
}
