% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jsonobj.R
\name{json_map}
\alias{json_map}
\title{json_map}
\usage{
json_map(...)
}
\arguments{
\item{...}{optional contents of the named list (as key-value pairs)}
}
\value{
a named list
}
\description{
Convenience function for making a (possibly empty) named list, which converts to a JSON object.
}
\examples{
json_map()
json_map(one = 1, two = TRUE, three = "THREE")
}
