% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_unique_name.R
\name{make_unique_eng_name}
\alias{make_unique_eng_name}
\title{make_unique_eng_name}
\usage{
make_unique_eng_name()
}
\value{
name in the form "Engine_###"
}
\description{
Convenience or internal function for generating engine names; each call results in a new name.
}
\examples{
make_unique_eng_name()
}

