% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_unique_name.R
\name{make_unique_rec_name}
\alias{make_unique_rec_name}
\title{make_unique_rec_name}
\usage{
make_unique_rec_name()
}
\value{
name in the form "Record_###"
}
\description{
Convenience or internal function for generating record names; each call results in a new name.
}
\examples{
make_unique_rec_name()
}

