% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_unique_name.R
\name{make_unique_symb_name}
\alias{make_unique_symb_name}
\title{make_unique_symb_name}
\usage{
make_unique_symb_name(symbols)
}
\arguments{
\item{symbols}{Symbols}
}
\value{
name in the form "tmp_###"
}
\description{
Convenience or internal function for generating symbol names; each call results in a new name.
}
\examples{
make_unique_symb_name()
}

