% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pfa.engine.R
\name{pfa.engine}
\alias{pfa.engine}
\title{pfa.engine}
\usage{
pfa.engine(config, tempFile = NULL)
}
\arguments{
\item{config}{list-of-lists representing a complete PFA document}

\item{tempFile}{if NULL, generate the PFA as a string in memory and pass it without touching disk; if a string, save the PFA document in a temporary file and have Python load it from that file}
}
\value{
an object with a $action(...) method that can be used to score data
}
\description{
Create an executable PFA scoring engine in R by calling Titus through rPython. If this function is successful, then the PFA is valid (only way to check PFA validity in R).
}
\examples{
pfaDocument1 <- pfa.config(avro.double, avro.double, expression(input + 10))
pfa.engine(pfaDocument1)   # where the pfaDocument is created by pfa.config
}

