% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pfa.gbm.R
\name{pfa.gbm.buildOneTree}
\alias{pfa.gbm.buildOneTree}
\title{pfa.gbm.buildOneTree}
\usage{
pfa.gbm.buildOneTree(tree, categoricalLookup, whichNode, valueNeedsTag = TRUE,
  dataLevels = NULL, fieldTypes = NULL)
}
\arguments{
\item{tree}{tree object}

\item{categoricalLookup}{splits used}

\item{whichNode}{left or right node for categoricalLookup}

\item{valueNeedsTag}{flag for whether node needs label}

\item{dataLevels}{levels of data}

\item{fieldTypes}{type of fields}
}
\value{
PFA as a list-of-lists that can be inserted into a cell or pool
}
\description{
Builds one tree extracted by pfa.gbm.extractTree.
}
\examples{
X1 <- runif(100)
X2 <- rnorm(100)
Y <- rexp(100,5) + 5 * X1 - 4 * X2
Y <- Y > 0
zz <- gbm(Y ~ X1 + X2)
zz$problemType <- "Classification"
zz1 <- pfa.gbm.extractTree(zz)
zz2 <- pfa.gbm.buildOneTree(zz1$treeTable, list(), 1)
}

