% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pfa.gbm.R
\name{pfa.gbm.extractTree}
\alias{pfa.gbm.extractTree}
\title{pfa.gbm.extractTree}
\usage{
pfa.gbm.extractTree(gbm, whichTree = 1)
}
\arguments{
\item{gbm}{an object of class "gbm"}

\item{whichTree}{FIXME}
}
\value{
FIXME
}
\description{
Extracts a tree from a forest made by the gbm library.
}
\examples{
FIXME
}

