% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbm.R
\name{pfa.gbm}
\alias{pfa.gbm}
\title{PFA Formatting of Fitted GBMs}
\source{
pfa_config.R avro_typemap.R avro.R pfa_cellpool.R pfa_expr.R
}
\usage{
\method{pfa}{gbm}(object, n.trees = NULL, name = NULL, version = NULL,
  doc = NULL, metadata = NULL, randseed = NULL, options = NULL, ...)
}
\arguments{
\item{object}{an object of class "gbm"}

\item{n.trees}{an integer or vector of integers specifying the number of trees 
to use in building the model. If a vector is provided, then only the indices of 
thos trees will be used. If a single integer is provided then all trees up until 
and including that index will be used.}

\item{name}{a character which is an optional name for the scoring engine}

\item{version}{an integer which is sequential version number for the model}

\item{doc}{a character which is documentation string for archival purposes}

\item{metadata}{a \code{list} of strings that is computer-readable documentation for 
archival purposes}

\item{randseed}{a integer which is a global seed used to generate all random 
numbers. Multiple scoring engines derived from the same PFA file have 
different seeds generated from the global one}

\item{options}{a \code{list} with value types depending on option name 
Initialization or runtime options to customize implementation 
(e.g. optimization switches). May be overridden or ignored by PFA consumer}

\item{...}{additional arguments affecting the PFA produced}
}
\value{
a \code{list} of lists that compose valid PFA document
}
\description{
This function takes a gradient boosted machine (gbm) fit using gbm
and returns a list-of-lists representing in valid PFA document 
that could be used for scoring
}
\examples{
X1 <- runif(100)
X2 <- factor(c(rep(c('A', 'B'), 30), 
               rep('C', 40)),
             ordered=F, 
             levels=c('B','C','A'))
Y <- ifelse(as.character(X2)=='B', 10, 20) * X1 - ifelse(as.character(X2)=='B', 15, 0) + rnorm(10,0,.5)

model <- gbm(Y ~ X1 + X2, n.trees = 2, interaction.depth=3, distribution = 'gaussian')
model_as_pfa <- pfa.gbm(model)

}
\seealso{
\code{\link[gbm]{gbm}}
}

