% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pfa.glm.R
\name{pfa.glm}
\alias{pfa.glm}
\title{PFA Formatting of Fitted GLMs}
\source{
pfa.config.R avro.typemap.R avro.R
}
\usage{
pfa(model)
}
\arguments{
\item{object}{an object of class "glm"}
}
\value{
a \code{list} of lists that compose valid PFA document
}
\description{
This function takes a generalized linear model fit using glm  
and returns a list-of-lists representing in valid PFA document 
that could be used for scoring
}
\examples{
X1 <- rnorm(100)
X2 <- runif(100)
Y <- 3 - 5 * X1 + 3 * X2 + rnorm(100, 0, 3)
Y <- Y > 0

model <- glm(Y ~ X1 + X2, family = binomial(logit))
model_as_pfa <- pfa(model)

}
\seealso{
\code{\link[stats]{glm}} \code{\link{pfa.glm.extractParams}}
}

