% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pfa.glmnet.R
\name{pfa.glmnet.extractParams}
\alias{pfa.glmnet.extractParams}
\title{pfa.glmnet.extractParams}
\usage{
pfa.glmnet.extractParams(cvfit, lambdaval = "lambda.1se")
}
\arguments{
\item{cvfit}{an object of class "cv.glmnet"}

\item{lambdaval}{FIXME}
}
\value{
PFA as a list-of-lists that can be inserted into a cell or pool
}
\description{
Extract generalized linear model net parameters from the glm library
}
\examples{
FIXME
}

