% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pfa.glmnet.R
\name{pfa.glmnet.inputType}
\alias{pfa.glmnet.inputType}
\title{pfa.glmnet.inputType}
\usage{
pfa.glmnet.inputType(params, name = NULL, namespace = NULL)
}
\arguments{
\item{params}{list of regressor names}

\item{name}{required name}

\item{namespace}{optional namespace}
}
\value{
List of regression parameter names
}
\description{
Extracts a set of regression variable names from a glmnet object
}
\examples{
xr <- list(regressors = c("x1", "x2"))
xrType <- pfa.glmnet.inputType(xr)
}

