% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pfa.glmnet.R
\name{pfa.glmnet.modelParams}
\alias{pfa.glmnet.modelParams}
\title{pfa.glmnet.modelParams}
\usage{
pfa.glmnet.modelParams(params)
}
\arguments{
\item{params}{function to extract glmnet model parameters}
}
\value{
glmnet intercept and coefficients
}
\description{
Extracts model coefficients
}
\examples{
x1 <- rnorm(100)
x2 <- runif(100)
X <- matrix(c(x1, x2), nc = 2)
Y <- 3 * x1 * x2
Y <- Y > 0
uu <- glmnet(X, Y)
vv <- pfa.glmnet.modelParams(list(const = uu$a0[1], coeff = uu$beta[1, ]))
}

