% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmnet.R
\name{pfa.glmnet}
\alias{pfa.glmnet}
\title{PFA Formatting of Fitted glmnet objects}
\source{
pfa_config.R avro_typemap.R avro.R pfa_cellpool.R pfa_expr.R
}
\usage{
\method{pfa}{glmnet}(object, lambda = NULL, name = NULL, version = NULL,
  doc = NULL, metadata = NULL, randseed = NULL, options = NULL, ...)
}
\arguments{
\item{object}{an object of class "glmnet"}

\item{lambda}{a numeric value of the penalty parameter lambda at which 
coefficients are required}

\item{name}{a character which is an optional name for the scoring engine}

\item{version}{an integer which is sequential version number for the model}

\item{doc}{a character which is documentation string for archival purposes}

\item{metadata}{a \code{list} of strings that is computer-readable documentation for 
archival purposes}

\item{randseed}{a integer which is a global seed used to generate all random 
numbers. Multiple scoring engines derived from the same PFA file have 
different seeds generated from the global one}

\item{options}{a \code{list} with value types depending on option name 
Initialization or runtime options to customize implementation 
(e.g. optimization switches). May be overridden or ignored by PFA consumer}

\item{...}{additional arguments affecting the PFA produced}
}
\value{
a \code{list} of lists that compose valid PFA document
}
\description{
This function takes a generalized linear model fit using glmnet  
and returns a list-of-lists representing a valid PFA document 
that could be used for scoring
}
\examples{
X <- matrix(c(rnorm(100), runif(100)), nrow=100, ncol=2)
Y <- factor(3 - 5 * X1 + 3 * X2 + rnorm(100, 0, 3) > 0)

model <- glmnet(X, Y, family = 'binomial')
model_as_pfa <- pfa(model)

}
\seealso{
\code{\link[glmnet]{glmnet}} \code{\link{extract_params.glmnet}}
}

