% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pfa.glmnet.R
\name{pfa.glmnet.regressionType}
\alias{pfa.glmnet.regressionType}
\title{pfa.glmnet.regressionType}
\usage{
pfa.glmnet.regressionType(params)
}
\arguments{
\item{params}{type of output variable}
}
\value{
type of output variable specified by params
}
\description{
Determines regression variable type
}
\examples{
pfa.glmnet.regressionType(list(binary = TRUE))
}

