% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pfa.lm.R
\name{pfa.lm}
\alias{pfa.lm}
\title{PFA Formatting of Fitted Linear models}
\source{
pfa.config.R avro.typemap.R avro.R
}
\usage{
\method{pfa}{lm}(object)
}
\arguments{
\item{object}{an object of class "lm"}
}
\value{
a \code{list} of lists that compose valid PFA document
}
\description{
This function takes a linear model fit using lm and 
returns a list-of-lists representing in valid PFA document 
that could be used for scoring
}
\examples{
X1 <- rnorm(100)
X2 <- runif(100)
Y <- 3 - 5 * X1 + 3 * X2 + rnorm(100, 0, 3)

model <- lm(Y ~ X1 + X2)
model_as_pfa <- pfa(model)

}
\seealso{
\code{\link[stats]{lm}} \code{\link{pfa.glm}}
}

