% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pfa.cellpool.R
\name{pfa.pool}
\alias{pfa.pool}
\title{pfa.pool}
\usage{
pfa.pool(type, init, source = "embedded", shared = FALSE,
  rollback = FALSE)
}
\arguments{
\item{type}{pool type, which is an Avro schema as list-of-lists (created by avro.* functions)}

\item{init}{pool initial value, which is a list-of-lists, usually converted from a model}

\item{source}{if "embedded", the init is the data structure, if "json", the init is a URL string pointing to an external JSON file}

\item{shared}{if TRUE, the pool is shared across scoring engine instances}

\item{rollback}{if TRUE, the pool's value would be rolled back if an uncaught exception is encountered}
}
\value{
a list-of-lists that can be inserted into pfa.config.
}
\description{
Creates a list-of-lists representing a PFA pool.
}
\examples{
pfa.pool(avro.double, json.map(one = 1.1, two = 2.2, three = 3.3))
}

