% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pfa.randomForest.R
\name{pfa.randomForest.extractTree}
\alias{pfa.randomForest.extractTree}
\title{pfa.randomForest.extractTree}
\usage{
pfa.randomForest.extractTree(forest, whichTree = 1, labelVar = FALSE)
}
\arguments{
\item{forest}{an object of class "randomForest"}

\item{whichTree}{which tree to extract}

\item{labelVar}{flag for whether var needs to be labeled}
}
\value{
single tree extracted from random forest
}
\description{
Extracts a tree from a forest made by the randomForest library.
}
\examples{
X1 <- runif(100)
X2 <- rnorm(100)
Y <- rexp(100,5) + 5 * X1 - 4 * X2
Y <- Y > 0
Y <- factor(Y)
zz <- randomForest(Y ~ X1 + X2)
zz1 <- pfa.randomForest.extractTree(zz)
}

