% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomForest.R
\name{pfa.randomForest}
\alias{pfa.randomForest}
\title{PFA Formatting of Fitted Random Forest Models}
\source{
pfa_config.R avro_typemap.R avro_R pfa_cellpool.R pfa_expr.R
}
\usage{
\method{pfa}{randomForest}(object, pred_type = c("response", "prob"),
  n.trees = NULL, name = NULL, version = NULL, doc = NULL,
  metadata = NULL, randseed = NULL, options = NULL, ...)
}
\arguments{
\item{object}{an object of class "randomForest"}

\item{n.trees}{an integer or vector of integers specifying the number of trees 
to use in building the model. If a vector is provided, then only the indices of 
thos trees will be used. If a single integer is provided then all trees up until 
and including that index will be used.}

\item{name}{a character which is an optional name for the scoring engine}

\item{version}{an integer which is sequential version number for the model}

\item{doc}{a character which is documentation string for archival purposes}

\item{metadata}{a \code{list} of strings that is computer-readable documentation for 
archival purposes}

\item{randseed}{a integer which is a global seed used to generate all random 
numbers. Multiple scoring engines derived from the same PFA file have 
different seeds generated from the global one}

\item{options}{a \code{list} with value types depending on option name 
Initialization or runtime options to customize implementation 
(e.g. optimization switches). May be overridden or ignored by PFA consumer}

\item{...}{additional arguments affecting the PFA produced}
}
\value{
a \code{list} of lists that compose valid PFA document
}
\description{
This function takes a random forest model fit using randomForest() and 
returns a list-of-lists representing in valid PFA document that could 
be used for scoring
}
\examples{
binomial_dat <- data.frame(X1 = runif(100), 
                           X2 = rnorm(100))
binomial_dat$Y <- factor((rexp(100,5) + 
                            5 * binomial_dat$X1 - 
                            4 * binomial_dat$X2) > 0)

model <- randomForest(Y ~ X1 + X2, data=binomial_dat)
model_as_pfa <- pfa.randomForest(model)
}
\seealso{
\code{\link[randomForest]{randomForest}}
}

