% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pfa.R
\name{pfa}
\alias{pfa}
\title{Generate PFA Document from Object}
\usage{
pfa(object, ...)
}
\arguments{
\item{object}{a model object for which a PFA document is desired}

\item{...}{additional arguments affecting the PFA produced.}
}
\value{
a \code{list} of lists that compose valid PFA document
}
\description{
pfa is a generic function for generating valid PFA documents from the results 
of various model fitting functions. The function invokes particular methods 
which depend on the class of the first argument.
}
\examples{
\dontrun{
require(utils)

# all the "pfa" methods found
methods("predict")
}
}
\seealso{
\code{\link{pfa.lm}} \code{\link{pfa.glm}}
}

