% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pfa_engine.R
\name{pfa_engine}
\alias{pfa_engine}
\title{pfa_engine}
\source{
json.R
}
\usage{
pfa_engine(config, temp_file = NULL)
}
\arguments{
\item{config}{\code{list} of lists representing a complete PFA document}

\item{temp_file}{if NULL, generate the PFA as a string in memory and pass it 
without touching disk; if a string, save the PFA document in a temporary file 
and have Python load it from that file}
}
\value{
an object with a $action(...) method that can be used to score data
}
\description{
Create an executable PFA scoring engine in R by calling Titus through rPython. 
If this function is successful, then the PFA is valid (only way to check 
PFA validity in R).
}
\examples{
my_pfa_doc <- pfa_document(avro_double, avro_double, expression(input + 10))
pfa_engine(my_pfa_doc)   # where the my_pfa_doc is created by pfa_document
}

