% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pfa_expr.R
\name{pfa_expr}
\alias{pfa_expr}
\title{pfa_expr}
\usage{
pfa_expr(expr, symbols = list(), cells = list(), pools = list(),
  fcns = list(), env = parent.frame())
}
\arguments{
\item{expr}{quoted R expression (e.g. quote(2 + 2))}

\item{symbols}{list of symbol names that would be in scope when evaluating this expression}

\item{cells}{list of cell names that would be in scope when evaluating this expression}

\item{pools}{list of pool names that would be in scope when evaluating this expression}

\item{fcns}{list of function names that would be in scope when evaluating this expression}

\item{env}{environment for resolving unrecognized symbols as substitutions}
}
\value{
a \code{list} of lists representing a fragment of a PFA document
}
\description{
Convert a quoted R expression into a \code{list} of lists that can be 
inserted into PFA
}
\examples{
pfa_expr(quote(2 + 2))
}
