% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pfa_cellpool.R
\name{pfa_pool}
\alias{pfa_pool}
\title{pfa_pool}
\usage{
pfa_pool(type, init, source = "embedded", shared = FALSE,
  rollback = FALSE)
}
\arguments{
\item{type}{pool type, which is an Avro schema as \code{list} of lists (created by avro_* functions)}

\item{init}{pool initial value, which is a \code{list} of lists, usually converted from a model}

\item{source}{if "embedded", the init is the data structure, if "json", the init is a URL string pointing to an external JSON file}

\item{shared}{if TRUE, the pool is shared across scoring engine instances}

\item{rollback}{if TRUE, the pool's value would be rolled back if an uncaught exception is encountered}
}
\value{
a \code{list} of lists that can be inserted into pfa_config.
}
\description{
Creates a \code{list} of lists representing a PFA pool.
}
\examples{
pfa_pool(avro_double, json_map(one = 1.1, two = 2.2, three = 3.3))
}
