% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_pfa.R
\name{read_pfa}
\alias{read_pfa}
\title{read_pfa}
\usage{
read_pfa(x)
}
\arguments{
\item{x}{A string or connection. A string can be either a path, a string of JSON.}
}
\value{
a \code{list} of lists structure in which null -> NULL, true -> TRUE, 
false -> FALSE, numbers -> numeric, strings -> character, array -> list, 
object -> named list
}
\description{
Convert a JSON string in memory or a JSON file on disk into a list-of-lists structure.
}
\examples{
\dontrun{
# literal JSON string  (useful for small examples)
toy_model <- read_pfa('{"input":"double","output":"double","action":[{"+":["input",10]}]}')

# from a local path
local_model <- read_pfa(system.file("extdata", "my-model.pfa", package = "aurelius"))

# from a url (split on two lines so not to exceed 100 char wide during install)
url_model <- read_pfa(paste0('https://raw.githubusercontent.com/ReportMort/hadrian', 
                     '/feature/add-r-package-structure/aurelius/inst/extdata/my-model.pfa'))
}
}

