% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniqueNames.R
\name{uniqueEngineName}
\alias{uniqueEngineName}
\title{uniqueEngineName}
\usage{
uniqueEngineName()
}
\value{
name in the form "Engine_###"
}
\description{
Convenience or internal function for generating engine names; each call results in a new name.
}
\examples{
uniqueEngineName()
}

