% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniqueNames.R
\name{uniqueEnumName}
\alias{uniqueEnumName}
\title{uniqueEnumName}
\usage{
uniqueEnumName()
}
\value{
name in the form "Enum_###"
}
\description{
Convenience or internal function for generating enum names; each call results in a new name.
}
\examples{
uniqueEnumName()
}

