% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/uniqueNames.R
\name{uniqueFixedName}
\alias{uniqueFixedName}
\title{uniqueFixedName}
\usage{
uniqueFixedName()
}
\value{
name in the form "Fixed_###"
}
\description{
Convenience or internal function for generating fixed names; each call results in a new name.
}
\examples{
uniqueFixedName()
}

