% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniqueNames.R
\name{uniqueRecordName}
\alias{uniqueRecordName}
\title{uniqueRecordName}
\usage{
uniqueRecordName()
}
\value{
name in the form "Record_###"
}
\description{
Convenience or internal function for generating record names; each call results in a new name.
}
\examples{
uniqueRecordName()
}

