% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/uniqueNames.R
\name{uniqueSymbolName}
\alias{uniqueSymbolName}
\title{uniqueSymbolName}
\usage{
uniqueSymbolName(symbols)
}
\value{
name in the form "tmp_###"
}
\description{
Convenience or internal function for generating symbol names; each call results in a new name.
}
\examples{
uniqueSymbolName()
}

