% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/avro.R
\docType{data}
\name{avro.boolean}
\alias{avro.boolean}
\title{avro.boolean}
\format{\preformatted{ chr "boolean"
}}
\usage{
avro.boolean
}
\description{
Constructs a list-of-lists Avro schema for the boolean (logical) type.
}
\keyword{datasets}

