% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/avro.R
\docType{data}
\name{avro.bytes}
\alias{avro.bytes}
\title{avro.bytes}
\format{\preformatted{ chr "bytes"
}}
\usage{
avro.bytes
}
\description{
Constructs a list-of-lists Avro schema for the bytes (unstructured byte array) type.
}
\keyword{datasets}

