% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/avro.R
\docType{data}
\name{avro.float}
\alias{avro.float}
\title{avro.float}
\format{\preformatted{ chr "float"
}}
\usage{
avro.float
}
\description{
Constructs a list-of-lists Avro schema for the float (floating-point numeric with 32-bit precision) type.
}
\keyword{datasets}

