% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/avro.R
\docType{data}
\name{avro.int}
\alias{avro.int}
\title{avro.int}
\format{\preformatted{ chr "int"
}}
\usage{
avro.int
}
\description{
Constructs a list-of-lists Avro schema for the int (integer numeric with 32-bit precision) type.
}
\keyword{datasets}

