% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avro.R
\docType{data}
\name{avro.null}
\alias{avro.null}
\title{avro.null}
\format{An object of class \code{character} of length 1.}
\usage{
avro.null
}
\description{
Constructs a list-of-lists Avro schema for the null type (type with only one value).
}
\keyword{datasets}

