% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avro.R
\docType{data}
\name{avro.string}
\alias{avro.string}
\title{avro.string}
\format{An object of class \code{character} of length 1.}
\usage{
avro.string
}
\description{
Constructs a list-of-lists Avro schema for the string (UTF-8) type.
}
\keyword{datasets}

