% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avro.typemap.R
\name{avro.typemap}
\alias{avro.typemap}
\title{avro.typemap}
\usage{
avro.typemap(...)
}
\arguments{
\item{...}{key-value pairs of Avro type schemas}
}
\value{
a function that yields Avro type schemas or just their names
}
\description{
Convenience function for ensuring that Avro type schemas are declared exactly once. It returns a function that yields a full type declaration the first time it is invoked and just a name on subsequent times.
}
\examples{
tm <- avro.typemap(
    MyType1 = avro.record(list(one = avro.int, two = avro.double, three = avro.string)),
    MyType2 = avro.array(avro.double)
)
tm("MyType1")           # produces the whole declaration
tm("MyType1")           # produces just "MyType1"
tm("MyType2")           # produces the whole declaration
tm("MyType2")           # produces the declaration again because this is not a named type
}

