% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbm.R
\name{build_node_gbm}
\alias{build_node_gbm}
\title{build_node_gbm}
\usage{
build_node_gbm(tree, categorical_lookup, leaf_val_type, whichNode,
  valueNeedsTag, dataLevels, fieldTypes)
}
\arguments{
\item{tree}{tree object}

\item{categorical_lookup}{splits used}

\item{leaf_val_type}{a character representing an avro type when a value is 
returned at a leaf node}

\item{whichNode}{left or right node for categorical_lookup}

\item{valueNeedsTag}{flag for whether node needs label}

\item{dataLevels}{levels of data}

\item{fieldTypes}{type of fields}
}
\value{
PFA as a list-of-lists that can be inserted into a cell or pool
}
\description{
Builds one node of a gbm model tree
}
\keyword{internal}
