% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pfa.gbm.R
\name{pfa.gbm.buildOneTree}
\alias{pfa.gbm.buildOneTree}
\title{pfa.gbm.buildOneTree}
\usage{
pfa.gbm.buildOneTree(tree, categoricalLookup, whichNode, valueNeedsTag = TRUE,
  dataLevels = NULL, fieldTypes = NULL)
}
\arguments{
\item{tree}{FIXME}

\item{categoricalLookup}{FIXME}

\item{whichNode}{FIXME}

\item{valueNeedsTag}{FIXME}

\item{dataLevels}{FIXME}

\item{fieldTypes}{FIXME}
}
\value{
FIXME
}
\description{
Builds one tree extracted by pfa.gbm.extractTree.
}
\examples{
FIXME
}

