% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbm.R
\name{pfa.gbm}
\alias{pfa.gbm}
\title{PFA Formatting of Fitted GBMs}
\source{
pfa_config.R avro_typemap.R avro.R pfa_cellpool.R pfa_expr.R pfa_utils.R
}
\usage{
\method{pfa}{gbm}(object, name = NULL, version = NULL, doc = NULL,
  metadata = NULL, randseed = NULL, options = NULL,
  pred_type = c("response", "prob"), cutoffs = NULL, n.trees = NULL, ...)
}
\arguments{
\item{object}{an object of class "gbm"}

\item{name}{a character which is an optional name for the scoring engine}

\item{version}{an integer which is sequential version number for the model}

\item{doc}{a character which is documentation string for archival purposes}

\item{metadata}{a \code{list} of strings that is computer-readable documentation for 
archival purposes}

\item{randseed}{a integer which is a global seed used to generate all random 
numbers. Multiple scoring engines derived from the same PFA file have 
different seeds generated from the global one}

\item{options}{a \code{list} with value types depending on option name 
Initialization or runtime options to customize implementation 
(e.g. optimization switches). May be overridden or ignored by PFA consumer}

\item{pred_type}{a string with value "response" for returning a prediction on the 
same scale as what was provided during modeling, or value "prob", which for classification 
problems returns the probability of each class.}

\item{cutoffs}{(Classification only) A named numeric vector of length equal to 
number of classes. The "winning" class for an observation is the one with the 
maximum ratio of predicted probability to its cutoff. The default cutoffs assume the 
same cutoff for each class that is 1/k where k is the number of classes}

\item{n.trees}{an integer or vector of integers specifying the number of trees 
to use in building the model. If a vector is provided, then only the indices of 
thos trees will be used. If a single integer is provided then all trees up until 
and including that index will be used.}

\item{...}{additional arguments affecting the PFA produced}
}
\value{
a \code{list} of lists that compose valid PFA document
}
\description{
This function takes a gradient boosted machine (gbm) fit using gbm
and returns a list-of-lists representing in valid PFA document 
that could be used for scoring
}
\examples{
dat <- data.frame(X1 = runif(100), 
                  X2 = rnorm(100))
dat$Y <- ((rexp(100,5) + 5 * dat$X1 - 4 * dat$X2) > 0)

bernoulli_model <- gbm::gbm(Y ~ X1 + X2, 
                            data = dat, 
                            distribution = 'bernoulli')
model_as_pfa <- pfa(bernoulli_model)
}
\seealso{
\code{\link[gbm]{gbm}}
}
