% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm.R
\name{pfa.glm.extractParams}
\alias{pfa.glm.extractParams}
\title{pfa.glm.extractParams}
\usage{
\method{pfa}{glm.extractParams}(fit, input_name = "glm_input",
  model_name = "glm_model")
}
\arguments{
\item{fit}{an object of class "glm"}

\item{input_name}{a character for the name of the input as it should appear 
in the PFA document}

\item{model_name}{a character for the name of the model as it should appear 
in the PFA document}
}
\value{
PFA as a list-of-lists that can be inserted into a cell or pool
}
\description{
Extract generalized linear model parameters from the glm library
}
\examples{
X1 <- rnorm(100)
X2 <- runif(100)
Y <- 3 - 5 * X1 + 3 * X2 + rnorm(100, 0, 3)
Y <- Y > 0

z <- glm(Y ~ X1 + X2, family = binomial(logit))
z2 <- pfa.glm.extractParams(z)
}

