% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pfa.glm.R
\name{pfa.glm.extractParams}
\alias{pfa.glm.extractParams}
\title{pfa.glm.extractParams}
\usage{
pfa.glm.extractParams(fit)
}
\arguments{
\item{fit}{an object of class "glm"}
}
\value{
PFA as a list-of-lists that can be inserted into a cell or pool
}
\description{
Extract generalized linear model parameters from the glm library
}
\examples{
X1 <- rnorm(100)
X2 <- runif(100)
Y <- 3 - 5 * X1 + 3 * X2 + rnorm(100, 0, 3)
Y <- Y > 0

z <- glm(Y ~ X1 + X2, family = binomial(logit))
z2 <- pfa.glm.extractParams(z)
}

