% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pfa.glm.R
\name{pfa.glm.extractParams}
\alias{pfa.glm.extractParams}
\title{pfa.glm.extractParams}
\usage{
pfa.glm.extractParams(fit)
}
\arguments{
\item{fit}{an object of class "glm"}
}
\value{
PFA as a list-of-lists that can be inserted into a cell or pool
}
\description{
Extract generalized linear model parameters from the glm library
}
\examples{
FIXME
}

