% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pfa.glmnet.R
\name{pfa.glmnet.extractParams}
\alias{pfa.glmnet.extractParams}
\title{pfa.glmnet.extractParams}
\usage{
pfa.glmnet.extractParams(cvfit, lambdaval = "lambda.1se")
}
\arguments{
\item{cvfit}{an object of class "cv.glmnet"}

\item{lambdaval}{value of tuning parameter}
}
\value{
PFA as a list-of-lists that can be inserted into a cell or pool
}
\description{
Extract generalized linear model net parameters from the glm library
}
\examples{
x1 <- rnorm(100)
x2 <- runif(100)
X <- matrix(c(x1, x2), nc = 2)
Y <- 3 * x1 * x2
hist(Y)
Y <- Y > 0
uu <- glmnet(X, Y)
}

