% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pfa.glmnet.R
\name{pfa.glmnet.modelParams}
\alias{pfa.glmnet.modelParams}
\title{pfa.glmnet.modelParams}
\usage{
pfa.glmnet.modelParams(params)
}
\arguments{
\item{x}{describeme}
}
\value{
describeme
}
\description{
describeme
}
\examples{
someExamples
}

