% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pfa.glmnet.R
\name{pfa.glmnet.predictProb}
\alias{pfa.glmnet.predictProb}
\title{pfa.glmnet.predictProb}
\usage{
pfa.glmnet.predictProb(params, input, model)
}
\arguments{
\item{x}{describeme}
}
\value{
describeme
}
\description{
describeme
}
\examples{
someExamples
}

