% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pfa.randomForest.R
\name{pfa.randomForest.buildOneTree}
\alias{pfa.randomForest.buildOneTree}
\title{pfa.randomForest.buildOneTree}
\usage{
pfa.randomForest.buildOneTree(tree, whichNode, valueNeedsTag, dataLevels,
  fieldTypes = NULL)
}
\arguments{
\item{tree}{FIXME}

\item{whichNode}{FIXME}

\item{valueNeedsTag}{FIXME}

\item{dataLevels}{FIXME}

\item{fieldTypes}{FIXME}
}
\value{
PFA as a list-of-lists that can be inserted into a cell or pool
}
\description{
Builds one tree extracted by pfa.randomForest.extractTree.
}
\examples{
FIXME
}

