% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pfa.randomForest.R
\name{pfa.randomForest.extractTree}
\alias{pfa.randomForest.extractTree}
\title{pfa.randomForest.extractTree}
\usage{
pfa.randomForest.extractTree(forest, whichTree = 1, labelVar = FALSE)
}
\arguments{
\item{forest}{an object of class "randomForest"}

\item{whichTree}{FIXME}

\item{labelVar}{FIXME}
}
\value{
FIXME
}
\description{
Extracts a tree from a forest made by the randomForest library.
}
\examples{
FIXME
}

