% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pfa_engine.R
\name{pfa_engine}
\alias{pfa_engine}
\title{pfa_engine}
\source{
json.R
}
\usage{
pfa_engine(doc)
}
\arguments{
\item{doc}{\code{list} of lists representing a complete PFA document}
}
\description{
Create an executable PFA scoring engine in R by calling Titus through rPython. 
If this function is successful, then the PFA is valid (only way to check 
PFA validity in R).
}
\examples{
\dontrun{
my_pfa_doc <- pfa_document(avro_double, avro_double, expression(input + 10))
pfa_engine(my_pfa_doc)   # requres rPython and Titus to be installed
}
}
