% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/converters.R
\name{unjson}
\alias{unjson}
\title{unjson}
\usage{
unjson(x)
}
\arguments{
\item{x}{input string or connection (if not open, this function opens it)}
}
\value{
a list-of-lists structure in which null -> NULL, true -> TRUE, false -> FALSE, numbers -> numeric, strings -> character, array -> list, object -> named list
}
\description{
Convert a JSON string in memory or a JSON file on disk into a list-of-lists structure.
}
\keyword{internal}
